+++ datacontract/cli.py
@@ -11,11 +11,10 @@ from rich.table import Table
 from typer.core import TyperGroup
 from typing_extensions import Annotated
 
-from datacontract.catalog.catalog import create_index_html, \
-    create_data_contract_html
+from datacontract.catalog.catalog import create_index_html, create_data_contract_html
 from datacontract.data_contract import DataContract
-from datacontract.init.download_datacontract_file import \
-    download_datacontract_file, FileExistsException
+from datacontract.init.download_datacontract_file import download_datacontract_file, FileExistsException
+from datacontract.publish.publish import publish_to_datamesh_manager
 
 console = Console()
 
@@ -163,7 +162,12 @@ class ExportFormat(str, Enum):
 @app.command()
 def export(
     format: Annotated[ExportFormat, typer.Option(help="The export format.")],
-    output: Annotated[Path, typer.Option(help="Specify the file path where the exported data will be saved. If no path is provided, the output will be printed to stdout.")] = None,
+    output: Annotated[
+        Path,
+        typer.Option(
+            help="Specify the file path where the exported data will be saved. If no path is provided, the output will be printed to stdout."
+        ),
+    ] = None,
     server: Annotated[str, typer.Option(help="The server name to export.")] = None,
     model: Annotated[
         str,
@@ -204,7 +208,7 @@ def export(
     if output is None:
         console.print(result, markup=False)
     else:
-        with output.open('w') as f:
+        with output.open("w") as f:
             f.write(result)
         console.print(f"Written result to {output}")
 
@@ -226,6 +230,20 @@ def import_(
     """
     result = DataContract().import_from_source(format, source)
     console.print(result.to_yaml())
+
+
+@app.command(name="publish")
+def publish(
+    location: Annotated[
+        str, typer.Argument(help="The location (url or path) of the data contract yaml.")
+    ] = "datacontract.yaml",
+):
+    """
+    Publish the data contract to the Data Mesh Manager
+    """
+    publish_to_datamesh_manager(
+        data_contract=DataContract(data_contract_file=location),
+    )
 
 
 @app.command(name="catalog")